First push since pulling Matt's master merge, added ignore sustain button in keymap and also added a sustain pedal button in main interface.  Seems to mostly work but behavior is sometimes inconsistent.

+++ bk_JUCE/bitKlavier/JuceLibraryCode/modules/juce_audio_basics/mpe/juce_MPEInstrument.cpp
@@ -491,7 +491,7 @@ void MPEInstrument::updateNoteTotalPitchbend (MPENote& note)
 {
     if (legacyMode.isEnabled)
     {
-        note.totalPitchbendInSemitones = note.pitchbend.asSignedFloat() * (float) legacyMode.pitchbendRange;
+        note.totalPitchbendInSemitones = note.pitchbend.asSignedFloat() * legacyMode.pitchbendRange;
     }
     else
     {
@@ -516,11 +516,11 @@ void MPEInstrument::updateNoteTotalPitchbend (MPENote& note)
         auto notePitchbendInSemitones = 0.0f;
 
         if (zone.isUsingChannelAsMemberChannel (note.midiChannel))
-            notePitchbendInSemitones = note.pitchbend.asSignedFloat() * (float) zone.perNotePitchbendRange;
+            notePitchbendInSemitones = note.pitchbend.asSignedFloat() * zone.perNotePitchbendRange;
 
         auto masterPitchbendInSemitones = pitchbendDimension.lastValueReceivedOnChannel[zone.getMasterChannel() - 1]
                                                             .asSignedFloat()
-                                          * (float) zone.masterPitchbendRange;
+                                          * zone.masterPitchbendRange;
 
         note.totalPitchbendInSemitones = notePitchbendInSemitones + masterPitchbendInSemitones;
     }
