+++ bk_JUCE/bitKlavier/JuceLibraryCode/modules/juce_core/native/juce_linux_SystemStats.cpp
@@ -209,7 +209,7 @@ int64 Time::getHighResolutionTicksPerSecond() noexcept
 
 double Time::getMillisecondCounterHiRes() noexcept
 {
-    return (double) getHighResolutionTicks() * 0.001;
+    return getHighResolutionTicks() * 0.001;
 }
 
 bool Time::setSystemTimeToThisTime() const
