+++ bk_JUCE/bitKlavier/JuceLibraryCode/modules/juce_gui_basics/mouse/juce_DragAndDropContainer.cpp
@@ -438,7 +438,7 @@ void DragAndDropContainer::startDragging (const var& sourceDescription,
                 if (distance > lo)
                 {
                     auto alpha = (distance > hi) ? 0
-                                                 : (float) (hi - distance) / (float) (hi - lo)
+                                                 : (hi - distance) / (float) (hi - lo)
                                                      + random.nextFloat() * 0.008f;
 
                     dragImage.multiplyAlphaAt (x, y, alpha);
